Look for fesetround in libm rather than the executable
We usually get lucky because GCC's runtime library includes this functions.
However, their proper place is in the math library and Clang's runtime
library does not include these functions on windows, so without this change
the function is not found.

Also bump openlibm to include the latest changes that make openlibm
work with clang on windows.