Use RLE on step hash to make arrays and ranges equal
Instead of hashing the values themselves, hash the first value and
differences between subsequent elements using run-length encoding.
This allows for 0(1) hashing of ranges consistent with AbstractArrays,
which means they can now compare equal.

Elements for which the - operator is not defined are hashed directly.
This assumes that types which can be used for ranges implement -.

[ci skip]